requirements.txt
absl-py==0.10.0
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.9
Adafruit-SSD1306==1.6.2
altgraph==0.17
astor==0.8.1
auto-py-to-exe==2.9.0
autopep8==1.5.4
beautifulsoup4==4.9.1
bitstring==3.1.9
blis==0.7.4
bottle==0.12.19
bottle-websocket==0.2.9
bs4==0.0.1
catalogue==2.0.4
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
chatbotAI==0.3.1.0
ChatterBot==1.0.2
chatterbot-corpus==1.2.0
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
comtypes==1.1.7
contextvars==2.4
cryptography==3.2.1
cycler==0.10.0
cymem==2.0.5
dataclasses==0.8
decorator==4.4.2
dlib==19.22.0
docopt==0.6.2
ecdsa==0.17.0
Eel==0.12.4
esptool==3.3.1
face-recognition==1.3.0
face-recognition-models==0.3.0
feedparser==6.0.2
future==0.18.2
futures==3.1.1
fuzzywuzzy==0.18.0
gast==0.2.2
geocoder==1.38.1
geographiclib==1.50
geojson==2.5.0
geopy==2.0.0
gevent==21.1.2
gevent-websocket==0.10.1
gnewsclient==1.12
google-pasta==0.2.0
goslate==1.5.1
greenlet==1.1.0
grpcio==1.32.0
gTTS==2.1.1
gTTS-token==1.1.3
h11==0.12.0
h5py==2.10.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
immutables==0.15
importlib-metadata==1.7.0
importlib-resources==5.1.2
inflect==5.0.2
instant-api-client==0.1.2
jaraco.itertools==5.0.0
Jinja2==2.11.3
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lightgbm==3.0.0
lxml==4.6.2
Markdown==3.2.2
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==3.3.3
more-itertools==8.6.0
MouseInfo==0.1.3
moviepy==1.0.3
mpg123==0.4
mpyg321==0.0.4
multitasking==0.0.9
murmurhash==1.0.5
nltk==3.5
numpy==1.19.3
opencv-python==4.4.0.42
openweathermap-cli==1.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.4
pathy==0.5.2
pefile==2019.4.18
Pillow==7.2.0
Pint==0.17
playsound==1.2.2
pocketsphinx==0.1.15
preshed==3.0.5
proglog==0.1.9
prompt-toolkit==3.0.8
protobuf==3.13.0
prsaw==0.3.1
PyAudio==0.2.11
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pydataset==0.2.0
PyDictionary==2.0.1
pygame==2.0.1
PyGetWindow==0.0.9
pyglet==1.5.13
pyinstaller==4.3
pyinstaller-hooks-contrib==2020.8
pymongo==3.11.3
PyMsgBox==1.0.9
pyOpenSSL==20.0.0
pyowm==2.10.0
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
pyserial==3.5
PySimpleGUI==4.18.0
python-dateutil==2.7.5
# python3-weather-api==0.0.3rc0
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.1
pyusb==1.2.1
pywhatkit==3.2
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==3.13
ratelim==0.1.6
reedsolo==1.5.4
regex==2020.7.14
requests==2.24.0
rfc3986==1.4.0
scikit-learn==0.23.2
scipy==1.5.2
selenium==3.141.0
sgmllib3k==1.0.0
six==1.15.0
smart-open==3.0.0
sniffio==1.2.0
soupsieve==2.0.1
spacy==3.0.6
spacy-legacy==3.0.5
SpeechRecognition==3.8.1
SQLAlchemy==1.2.19
srsly==2.4.1
tabulate==0.8.7
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
textblob==0.15.3
tflearn==0.3.2
thinc==8.0.3
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.49.0
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.25.10
wasabi==0.8.2
wcwidth==0.2.5
weather-application3==0.3
Werkzeug==1.0.1
whichcraft==0.6.1
wikipedia==1.4.0
wincertstore==0.2
wit==6.0.0
wolframalpha==4.0.0
wrapt==1.12.1
XlsxWriter==1.3.6
xmltodict==0.12.0
yfinance==0.1.55
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.4.0
